@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows10.0.22000.0</TargetFramework>
+    <TargetFramework>net7.0-windows</TargetFramework>
     <LangVersion>latest</LangVersion>
     <UseWPF>true</UseWPF>
     <GenerateAssemblyTitleAttribute>false</GenerateAssemblyTitleAttribute>
@@ -47,24 +47,20 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Resource Include="..\pic\wechat.jpg" Link="Resources\wechat.jpg" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="ARSoft.Tools.Net" Version="2.3.0" />
-    <PackageReference Include="CryptoBase" Version="1.7.1" />
-    <PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="1.1.0" />
-    <PackageReference Include="HMBSbige.SingleInstance" Version="6.0.0" />
-    <PackageReference Include="HMBSbige.UpdateChecker" Version="6.0.1" />
+    <PackageReference Include="CryptoBase" Version="1.7.2" />
+    <PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="2.0.1" />
+    <PackageReference Include="HMBSbige.SingleInstance" Version="6.0.1" />
+    <PackageReference Include="HMBSbige.UpdateChecker" Version="8.0.1" />
     <PackageReference Include="RouteMatcher" Version="1.0.0" />
     <PackageReference Include="RunAtStartup" Version="5.0.2" />
-    <PackageReference Include="Shadowsocks.Protocol" Version="1.0.0" />
-    <PackageReference Include="Syncfusion.SfGrid.WPF" Version="19.4.0.43" />
-    <PackageReference Include="Syncfusion.SfTreeView.WPF" Version="19.4.0.43" />
+    <PackageReference Include="Shadowsocks.Protocol" Version="1.0.2" />
+    <PackageReference Include="Syncfusion.SfGrid.WPF" Version="28.1.39" />
+    <PackageReference Include="Syncfusion.SfTreeView.WPF" Version="28.1.39" />
     <PackageReference Include="URIScheme" Version="5.0.0" />
-    <PackageReference Include="WindowsProxy" Version="5.0.6" />
+    <PackageReference Include="WindowsProxy" Version="6.0.0" />
     <PackageReference Include="WpfColorFontDialog" Version="1.0.8" />
-    <PackageReference Include="ZXing.Net.Bindings.Windows.Compatibility" Version="0.16.10" />
+    <PackageReference Include="ZXing.Net.Bindings.Windows.Compatibility" Version="0.16.12" />
   </ItemGroup>
 
   <ItemGroup>
