@@ -1,18 +1,12 @@
 using System;
 using System.Security.Cryptography;
 
-namespace Shadowsocks.Util
-{
-    public static class Rng
-    {
-        public static void RandBytes(byte[] buf, int length = -1)
-        {
-            RandomNumberGenerator.Fill(length < 0 ? buf : buf.AsSpan(0, length));
-        }
+namespace Shadowsocks.Util;
 
-        public static string RandId()
-        {
-            return Guid.NewGuid().ToString().Replace(@"-", string.Empty);
-        }
+public static class Rng
+{
+    public static void RandBytes(byte[] buf, int length = -1)
+    {
+        RandomNumberGenerator.Fill(length < 0 ? buf : buf.AsSpan(0, length));
     }
 }
